refactor: extract FooterGradient component and optimize scroll/render logic
- Extract FooterGradient as memo'd component inside dark theme context
- Move scroll-to logic from individual benefit items to parent wrapper
- Convert renderLoginPrompt from IIFE to useMemo for render stability
